Apply Weka modifications.
In future, branch from weka-master when ldc makes a release branch. But base future development on weka-master. When something is reverted in an ldc release branch, but not in ldc master, we need to base newer weka version on master, not on the weka version based on an ldc release branch.