gha: don't cross-compile runtime to linux aarch64
The only thing we need to make it work is working cross-clang.
And the one we use can actually cross-compile to aarch64, but
it doesn't have rtlib for aarch64. And apparently having everything
set up properly on GHA is a PITA.

So, disable it for now and wait till we bump the compiler, so
we could build linux aarch64 natively and take runtime from it.